requirements.txt
pandas
yfinance
sqlalchemy
psycopg2-binary
python-dotenv
tenacity
pydantic 